refactor: move ecstore services domain modules
