+++ crates/ecstore/src/services/mod.rs
@@ -16,3 +16,5 @@ pub(crate) mod batch_processor;
 pub(crate) mod event_notification;
 pub(crate) mod metrics_realtime;
 pub(crate) mod notification_sys;
+pub(crate) mod rebalance;
+pub(crate) mod tier;
