+++ crates/ecstore/src/services/notification_sys.rs
@@ -16,9 +16,9 @@ use crate::cluster::rpc::PeerRestClient;
 use crate::diagnostics::admin_server_info::get_commit_id;
 use crate::error::{Error, Result};
 use crate::layout::endpoints::EndpointServerPools;
-use crate::rebalance::RebalSaveOpt;
 use crate::runtime::sources as runtime_sources;
 use crate::services::metrics_realtime::{CollectMetricsOpts, MetricType};
+use crate::services::rebalance::RebalSaveOpt;
 use crate::storage_api_contracts::admin::StorageAdminApi;
 use futures::future::join_all;
 use lazy_static::lazy_static;
