+++ crates/ecstore/src/services/rebalance/entry.rs
