+++ crates/ecstore/src/services/rebalance/worker.rs
