+++ crates/ecstore/src/services/tier/tier_gen.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::tier::tier::TierConfigMgr;
+use crate::services::tier::tier::TierConfigMgr;
 
 #[allow(dead_code)]
 impl TierConfigMgr {
