+++ crates/ecstore/src/services/tier/warm_backend_s3sdk.rs
@@ -35,7 +35,7 @@ use crate::client::{
 };
 use crate::error::ErrorResponse;
 use crate::error::error_resp_to_object_err;
-use crate::tier::{
+use crate::services::tier::{
     tier_config::TierS3,
     warm_backend::{WarmBackend, WarmBackendGetOpts},
 };
